Throw error if one has a few refs to modifying CTE
Multiple references to modifying CTEs can lead to DML operations being
dublicated. This commit proposes the restriction, in case of non-fulfillment
of which the informative error is thrown. The if-clause, which checks whether
current cte is referenced more than once and subplan sharing is disabled,
is proposed. If it is true, the error is thrown. This solution should be
replaced with more complex logic in future, providing the proper
materialization and rescan of modifying CTE.

Co-Authored-By: Alexey Gordeev <goa@arenadata.io>