ORCA generates incorrect plan with exists clause for partitioned table
For subqueries with EXISTS, the first column is always selected by ORCA.
But if this column does not participate anywhere else in the query and
is neither a system nor a distribution key, then it is not marked as used.
Therefore, when translating to DXL, such a column is skipped, and
in a debug build, such a plan is generally not allowed!
On a release build, when a column is skipped,
that column is not added to table_descr.
And therefore, its number is not inserted into the m_colid_to_attno_map
hashmap of the correspondence of the column number to the attribute number.
And now the attribute number is not found by the column number.
And so a variable with an unknown attribute number is created.
Such a variable with an unknown attribute number is treated as a whole-row.
In this case, the oid of this variable does not
correspond to the composite type and an error occurs.

The solution to the problem is very simple -
always mark the selected column as used.