Fix unavailability of all segment files after vacuum
After compaction the vacuum query always marked the segment files as awaiting
drop in the coordinator's hashmap, regardless of their actual state on segments.
Because the compaction might not have happened, the segment files might have
remained in their normal state, but coordinator did not get information about
this. Drop phase may be skipped in the case of there are concurrent
transactions, so segment files were left in awaiting drop state in
the coordinator's hashmap. Therefore on intensive concurrent insertion into
a table, vacuum marked all available segment files as awaiting drop in
the coordinator's hashmap and the next insert failed with message:

ERROR: could not find segment file to use for inserting into relation

After the patch the coordinator extracts information from the segments about
the actual state of the segment files after compaction and marks as awaiting
drop only those segment files that are actually compacted. This allows for
intensive concurrent insertion into the table.