Orca memory pool refactoring
Declare methods as static if they work with static member variables only

- I made the CMemoryPoolManager::Cleanup method static,
which was previously called only on the static object m_memory_pool_mgr.

- I made the CMemoryPoolManager::CreateMemoryPool method static,
which was previously called only on the static object m_memory_pool_mgr.

- I made the CMemoryPoolManager::Destroy method static,
which was previously called only on the static object m_memory_pool_mgr.

- I made the CMemoryPoolManager::GetGlobalMemoryPool method static,
which was previously called only on the static object m_memory_pool_mgr.

- I made the CMemoryPoolManager::Shutdown method static,
which was previously called only on the static object m_memory_pool_mgr.

- I made the CMessageRepository::Shutdown method static,
which was previously called on the static object m_repository.

I added assertions to make sure the pointers being dereferenced are not null.

This is backport of 60c1cf578d21c41500bb1cf4c76c2126230df8ea with additions:

- I made the CCacheFactory::Shutdown method static
  because it only uses static variables.

- I made the CFSimulator::Shutdown method static,
  which was previously called only on the static object m_fsim.

- I made the CWorkerPoolManager::Shutdown method static
  because it uses only static variables.

- I made the CXformFactory::Shutdown method static
  because it uses only static variables.