Fix wrong slice allocation for shared CTEs with modifying DML inside.
When the shared CTE with modifying DML inside is planned, at the first
occurrence of the CTE create_ctescan_plan function constructs the subplan, so
it's slice gets the gang type GANGTYPE_PRIMARY_WRITER. For the other occurrences
the subplan construction is being skipped, so if they are located in other
slices, and the CTE is modifying, their slices may not have
GANGTYPE_PRIMARY_WRITER gang type. Later, the ShareInputScans are assigned the
producer and consumer roles by apply_shareinput_dag_to_tree function, which
uses tree walker with specific traverse order. There was a case when the
consumer was created at GANGTYPE_PRIMARY_WRITER slice, while the producer were
put to the GANGTYPE_PRIMARY_READER slice. This lead to execution error when the
reader gang attempted to execute ModifyTable node.

This patch fixed the issue the following way. When creating the shared plan at
the first time, the gangType of root->curSlice is checked. If it has switched
compared to the value before the planning, we mark ShareInputScan as writer
slice maker (the rootSliceIsWriter flag is set to true). At other occurences of
CTE their ShareInputScan nodes are also marked with this state. After the
consumer, producer roles are assigned to each ShareInputScan, we set the
correct gang type to the producer (GANGTYPE_PRIMARY_WRITER), and to the
consumers (GANGTYPE_PRIMARY_READER, if the consumer is in different slice from
producer) in shareinput_mutator_xslice_2() function.