Fix orphaned file descriptors
Problem:
In cases where an AO table is created, some data is inserted, and then the
table is dropped, file descriptor of segfile will be opened still. It
stays open until session ends. Thus, disc space is not freed until the end
of the session.

Cause:
There are 2 issues:
1) mdcreate_ao is used to create a segment file during data insertion.
For creation, it calls PathNameOpenFile without closing.
File descriptor returned by PathNameOpenFile in mdcreate_ao is not stored
anywhere except the virtual file level (fd.c), where it will be cleaned
some time later only when the number of opened descriptors reaches threshold.
2) second issue happens when we run out of space. This problem is related to
file descriptors left open because of an abnormal abort of the insert command.
In function BufferedAppendWrite, once free space has ended, ereport is thrown.
Before that, the file is not closed, so after the long jump it remained open.

Fix:
Add FileClose to mdcreate_ao and BufferedAppendWrite (before throwing ereport).