Fixed docker files for running the gpdb7 test suite
The container base has been changed from CentOS 7 to Rocky 8.8.
Dnf is now used instead of yum as a package manager.
Devtoolset-7 was removed since the necessary tools are available in the main
repository.
Added installation of llvm and clang for building with JIT support.
Packages for llvm and clang are taken from the vault, since only clang-16 is
available in the main repository, and the build gpdb with llvm 16+ is not
supported.
RSA1 key generation was removed since it's deprecated.
The mdw container was renamed to cdw.
Now, instead of a combination of bash and tail, the init process and sleep
infinity are used in containers to prevent zombie processes from appearing.
In the README updated the docker commands and removed outdated information.