Add check for "rule->parname is not NULL" before using (#698)
The PartitionRule structure can be allocated in the add_part_to_catalog
function, where the parname argument may be NULL. It means, that the parname
field at this structure will not be set. When this structure will be processed
at the get_partition_recursive function, this filed will be used without
checking. It can lead to segfault.