Add check for "parentPolicy is not NULL" before using (#700)
Pointer to the parentPolicy variable is set from pointer
"parentrel->rd_cdbpolicy". This variable is checked before using at other
places at the transformDistributedBy function in case of the IsBinaryUpgrade
flag is false. At the last "if" the parentPolicy variable is used without
checking. The parentPolicy variable can be NULL for binary upgrade. For binary
upgrade this check runs against a segment in utility mode and the distribution
policy isn't stored in the segments.

The parentPolicy pointer is used at the make_distributedby_for_rel function.
We should execute this function only if the parentPolicy pointer
("parentrel->rd_cdbpolicy") is not NULL.