Remove dependency to system calls for memory allocation in refint and regress_gp (#784)
Failures in allocations could lead to crashes with NULL pointer
dereferences . Memory context TopMemoryContext is used instead to keep
alive the plans allocated in the session. A more specific context could
be used here, but this is left for later.

This is backport from postgres/b0b61963866 commit for refint file.
Also logic of this commit was applied to regress_gp file.