Treat missing prototypes as errors
This patch makes gcc treat missing prototypes as errors. Also, it fixes custom
allocators for zstd and autoconf file adding HAVE_LIBZSTD flag. This flag was
there initially and was generated automatically when linking zstd.so, but since
we moved to static linking HAVE_LIBZSTD changed to HAVE_LIB_LIBZSTD which is
incorrect