Fix panic with 'stuck spinlock' during crash recovery tests
Problem:
Isolation crash recovery tests (for example, uao_crash_compaction_column,
crash_recovery_dtm, etc) may fail with 'stuck spinlock' panic with very low
reproduction rate.

Cause:
Problem happens in the following case:
1. During the test a fault is injected from the test script, backend processes
it and goes into a panic (expected).
2. At the very same moment, background writer process calls
'SIMPLE_FAULT_INJECTOR("fault_in_background_writer_main")', he does it regularly
from the loop in 'BackgroundWriterMain'. Bg writer acquires the spinlock inside
the fault injector, but before it releases it...
3. Postmaster starts killing all child processes by sending SIGQUIT signal.
4. Background writer process halts its execution without releasing the spinlock.
5. In the background writer process handler for the SIGQUIT signal is invoked -
'bg_quickdie'. And 'bg_quickdie' calls
'SIMPLE_FAULT_INJECTOR("fault_in_background_writer_quickdie");'. As the spinlock
is not released, 'bg_quickdie' hangs on the spinlock.
6. Postmaster fails to wait child processes completion. And after timeout tries
to check 'SIMPLE_FAULT_INJECTOR("postmaster_server_loop_no_sigkill")' and hangs
on the same spinlock as well.
7. Finally both Postmaster and 'bg_quickdie' fails with 'stuck spinlock' panic.

Fix:
The general rule would be that it is NOT allowed to access the fault injector
from Postmaster process or from any SIGQUIT handler of child processes when the
system is resetting after the crash of some backend. Reason: during reset
Postmaster kills child processes, and it might kill some process when the
process has acquired the spin lock of the fault injector, but hasn't released it
yet. So, subsequent calls to the fault injector api from Postmaster or any
SIGQUIT handler will lead to deadlock. Only 'doomed' processes can still call
the fault injector api, as they will soon be terminated anyway. Solution is
following:
1. Remove 'SIMPLE_FAULT_INJECTOR("fault_in_background_writer_quickdie");' from
'bg_quickdie'.
2. Remove 'SIMPLE_FAULT_INJECTOR("postmaster_server_loop_no_sigkill")' from
Postmaster's ServerLoop.
3. As these two faults were used by the fts_segment_reset test, this test and
related code were redesigned. Now, instead of sleeping for a delay defined by
the test in 'bg_quickdie', Postmaster will send SIGSTOP to the bg writer, and
start a timer for the delay. Once the delay elapses, the timer handler will send
SIGCONT and SIGQUIT to the bg writer. This logic is triggered if the Postmaster
detects newly added fault "postmaster_delay_termination_bg_writer" (which is a
replacement for "fault_in_background_writer_quickdie" and
"postmaster_server_loop_no_sigkill"). This fault is checked only when Postmaster
is in the PM_RUN state. So it should be safe to check for it.
4. Also, 'SIMPLE_FAULT_INJECTOR_WITH_PARAM' was added to get the extra argument
set by the test. It is used to pass the delay value from the test.