Avoid linking AO aux tables to extension
Unlike TOAST tables, AO aux tables are linked to extension during its creation.
When table is dropped from extension using ALTER, aux tables are left as
extension dependencies. It means we can't DROP a table after that, because AO
aux tables are still used by extension. In other words, this behavior makes
impossible to DROP AO table that was created by extension without extension
DROP.

The goal of the patch is to change this behavior, allowing to DROP a table.
Unnecessary dependencies are now not created in heap_create_with_catalog(). The
behavior is the same as for TOAST tables, which are aux tables too. The
dependency created in CreateAOAuxiliaryTable() is enough.

The patch contains a test that describes target behavior. However, the test can
be run only manually. GPDB is not ready to run tests in modules subdir
automatically.