Fix segfault in get_ao_compression_ratio
Problem:
Segfault happened in some queries with get_ao_compression_ratio in target list.

Cause:
Function 'get_ao_compression_ratio' executes a query on
'pg_aoseg.pg_aoseg_<oid>' table to calculate compression ratio. A function that
accesses hash-distributed or randomly distributed tables must be executed on
coordinator. But planner created plans with 'get_ao_compression_ratio' execution
on segments. Inside 'get_ao_compression_ratio' there was an assert to check this
case, but on release build the assert was removed. So, the code of the function
was executed on a segment and got unexpected results from
'pg_aoseg.pg_aoseg_<oid>', that led to segfault.

Fix:
Assert is changed to runtime check. Now, if the planner creates a plan with the
function execution on a segment, user will get a proper error message. It is
also done for similar functions 'gp_update_ao_master_stats' and
'get_ao_distribution'. Plus, some unreachable code is removed.

Note: it seems that a fully proper way would be to mark these functions as
PROEXECLOCATION_MASTER instead of PROEXECLOCATION_ANY, but that would require
changing of system catalog, which is acceptable only between major versions
update. So this approach is not used.