Fix incorrect file descriptor dup when creating a pipe
If fcntl fails with F_DUPFD_CLOEXEC, attempts are made to duplicate the file
descriptor using a regular dup. But since, as a result of fcntl failure,
write_fd will become -1 and dup will also fail.

This patch returns write_fd to its original value in case of fcntl failure.